added card component, also logic for rendering 'new initiative' UI
