+++ ts/features/idpay/wallet/screens/BonusDetailsScreen.tsx
@@ -1,6 +1,6 @@
-import { View } from "native-base";
+import { Text, View } from "native-base";
 import React from "react";
-import { ImageBackground, ScrollView, StyleSheet } from "react-native";
+import { ScrollView, StyleSheet } from "react-native";
 import LinearGradient from "react-native-linear-gradient";
 import { widthPercentageToDP } from "react-native-responsive-screen";
 import { SafeAreaView } from "react-native-safe-area-context";
@@ -12,47 +12,36 @@ import { IOColors } from "../../../../components/core/variables/IOColors";
 import { IOStyles } from "../../../../components/core/variables/IOStyles";
 import BaseScreenComponent from "../../../../components/screens/BaseScreenComponent";
 import FocusAwareStatusBar from "../../../../components/ui/FocusAwareStatusBar";
-import cardBg from "../../../../../img/bonus/cgn/card_mask.png";
-import BonusCardComponent from "../components/BonusCard";
+import FooterWithButtons from "../../../../components/ui/FooterWithButtons";
+import customVariables from "../../../../theme/variables";
+import BonusCardComponent from "../components/BonusCardComponent";
 
 const styles = StyleSheet.create({
   card: {
     position: "absolute",
     alignSelf: "center",
-    width: widthPercentageToDP(83),
-    maxWidth: 340,
-    height: 168,
+    width: widthPercentageToDP(90),
+    maxWidth: 343,
+    height: 192,
     top: 2
-    // backgroundColor: "red"
   }
-  //   cardContainer: {
-  //     height: "100%",
-  //     width: widthPercentageToDP(90),
-  //     maxWidth: 340
-  //   },
-  //   imageFull: {
-  //     resizeMode: "stretch",
-  //     height: 215,
-  //     width: widthPercentageToDP(95),
-  //     maxWidth: 360,
-  //     top: -5,
-  //     left: -10,
-  //     zIndex: 8
-  //   }
 });
 
-export const BonusDetailsScreen = (props: Props) => {
+// const
+
+export const BonusDetailsScreen = () => {
   const bonus: InitiativeDTO = {
-    nInstr: "1",
+    nInstr: "0",
     endDate: new Date("2021-12-31"),
     initiativeId: "1",
     iban: "IT60X0542811101000000123456",
-    status: StatusEnum.REFUNDABLE,
+    status: StatusEnum.NOT_REFUNDABLE,
     accrued: 0,
     available: 500,
     refunded: 0,
-    initiativeName: "Circuito del Vino"
+    initiativeName: "18App"
   };
+
   return (
     <BaseScreenComponent
       dark={true}
@@ -61,7 +50,10 @@ export const BonusDetailsScreen = (props: Props) => {
       headerTitle={"Bonus"}
       headerBackgroundColor={IOColors.bluegrey}
     >
-      <FocusAwareStatusBar backgroundColor={IOColors.black} />
+      <FocusAwareStatusBar
+        backgroundColor={IOColors.bluegrey}
+        barStyle={"light-content"}
+      />
       <ScrollView style={IOStyles.flex} bounces={false}>
         <LinearGradient colors={[IOColors.bluegrey, IOColors.bluegreyDark]}>
           <View style={[IOStyles.horizontalContentPadding, { height: 149 }]} />
@@ -69,7 +61,29 @@ export const BonusDetailsScreen = (props: Props) => {
         <View style={styles.card}>
           <BonusCardComponent {...bonus} />
         </View>
+
+        <View
+          style={[
+            IOStyles.flex,
+            IOStyles.horizontalContentPadding,
+            { paddingTop: customVariables.contentPadding }
+          ]}
+        >
+          <View spacer extralarge />
+          <View spacer small />
+          {bonus.nInstr === "0" && <Text>INIZIATIVA NON CONFIGURATA</Text>}
+          
+        </View>
       </ScrollView>
+      {bonus.nInstr === "0" && (
+        <>
+          <FooterWithButtons
+            type="SingleButton"
+            leftButton={{ block: true, primary: true, title: "Inizia" }}
+          />
+          <View spacer />
+        </>
+      )}
     </BaseScreenComponent>
   );
 };
